address generation refactoring
